+++ src/main/java/com/raul/negocio/LogicaNegocio.java
@@ -14,21 +14,21 @@ import com.raul.datos.ConsultaDatos;
  */
 public class LogicaNegocio {
 	
-	public  int calculaDescuento(String nombre, String edad) throws IOException {
+	public  double calculaDescuento(String nombre, String edad) throws IOException {
 		//Usar clase datos
 		ConsultaDatos c = new ConsultaDatos();
-		Integer precio;
-		Integer descuento;
+		double precio;
+		double descuento;
 		int anos = Integer.parseInt(edad);
 		if (anos<20) {
-			descuento = (int) (c.consultaPrecio(nombre)* 0.2);
+			descuento = (double) (c.consultaPrecio(nombre)* 0.2);
 			precio=c.consultaPrecio(nombre)-descuento;
 		}else if((anos>=20) && (anos <=30)){
-			descuento = (int) (c.consultaPrecio(nombre)* 0.1);
+			descuento = (double) (c.consultaPrecio(nombre)* 0.1);
 			precio=c.consultaPrecio(nombre)-descuento;
 		}else {
-			descuento =(int) (c.consultaPrecio(nombre)* 0.05);
-			precio=c.consultaPrecio(nombre)-descuento;
+			descuento =(double) (c.consultaPrecio(nombre)* 0.05);
+			precio = c.consultaPrecio(nombre)-descuento;
 		}
